feat: increase max request size
This commit increases the maximum Send and Recieve request sizes in the
API server and the etcd client.

This allows to process and store objects larger than the the default K8s
and etcd limits (in the ballpark of 2 MiB).

Since TransportConfig does not expose request limits for etcd, the
easies way to change the etcd client limits is vendoring the module and
editing the etcd client factory function.

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>