build(makefile): fix target name conflict with "build" directory
Prior to this commit, introducing a "build" directory broke building a
standalone binary using a Makefile. Because the "build" directory exists
in the source tree, it made `make` think that the target already exists
and is up-to-date, so it would not even try to build new binaries.

This commit changes the target name and fixes the problem.

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>